+++ app/Configuration.hs
@@ -24,20 +24,28 @@ module Configuration
   where
 
 import Control.Applicative ((<|>), pure)
+import Control.Monad (when)
 import Data.Bool (Bool(False))
 import Data.Eq (Eq)
-import Data.Foldable (concatMap)
-import Data.Function ((.))
+import Data.Function (($), (.))
 import Data.Functor ((<$>))
 import Data.List.NonEmpty (NonEmpty((:|)))
-import Data.Maybe (Maybe(Just, Nothing), maybe)
+import Data.Maybe (Maybe(Nothing), maybe)
+import Data.Ord ((>=))
 import Data.Semigroup ((<>))
 import Data.String (String)
 import GHC.Generics (Generic)
-import System.IO (FilePath, Handle, IO, hIsTerminalDevice)
+import System.Exit (exitFailure)
+import System.IO
+    ( FilePath
+    , Handle
+    , IO
+    , hIsTerminalDevice
+    , hPutStrLn
+    , stderr
+    )
 import Text.Show (Show)
 
-import qualified Data.LanguageCodes as LanguageCode (ISO639_1(EN))
 import Data.Output.Colour
     ( ColourOutput
     , terminalSupportsColours
@@ -45,6 +53,7 @@ import Data.Output.Colour
     )
 import qualified Data.Output.Colour as ColourOutput (ColourOutput(Auto))
 import Data.Text (Text)
+import qualified Data.Text as Text (unpack)
 import Data.Verbosity (Verbosity)
 import qualified Data.Verbosity as Verbosity (Verbosity(Normal))
 import Dhall (FromDhall)
@@ -61,8 +70,9 @@ import qualified Dhall
     )
 import System.Console.Terminfo (setupTermFromEnv)
 import System.Directory (XdgDirectory(XdgCache), getXdgDirectory)
+import System.Environment.Parser (ParseEnvError(..), parseEnvIO)
 
-import TldrClient.Locale (Locale(Locale, country, language), decodeLocale, lookupLang)
+import TldrClient.Locale (Locale, decodeLocale, languagePreference)
 
 
 data Configuration = Configuration
@@ -140,26 +150,27 @@ getCacheDirectory :: Configuration -> IO FilePath
 getCacheDirectory Configuration{cacheDirectory} =
     maybe (getXdgDirectory XdgCache "tldr-client") pure cacheDirectory
 
--- TODO: This should be returning NonEmpty Locale.
-getLocales :: Configuration -> Maybe Locale -> IO [Locale]
-getLocales Configuration{verbosity, locale} override = do
-    -- TODO: This doesn't follow the specification yet!
-    ls <- maybe (pure <$> lookupLang verbosity defaultLocale) (pure . pure)
+getLocales :: Configuration -> Maybe Locale -> IO (NonEmpty Locale)
+getLocales Configuration{verbosity, locale} override =
+    maybe (parseEnvIO () dieParseEnvError languagePreference) (pure . pure)
         (override <|> locale)
-    pure (concatMap @[] generaliseLocale ls)
   where
-    defaultLocale :: Locale
-    defaultLocale = Locale
-        { language = LanguageCode.EN
-        , country = Nothing
-        }
-
-    generaliseLocale :: Locale -> [Locale]
-    generaliseLocale = \case
-        l@Locale{country = Just _} ->
-            [l, l{country = Nothing}]
-        l ->
-            [l]
+    dieParseEnvError :: ParseEnvError -> IO a
+    dieParseEnvError err = do
+        when (verbosity >= Verbosity.Normal) do
+            hPutStrLn stderr $ "Error: " <> case err of
+                ParseEnvError var msg ->
+                    Text.unpack var
+                    <> ": Failed to parse environment variable: "
+                    <> msg
+                MissingEnvVarError var ->
+                    Text.unpack var
+                    <> ": Missing required environment variable."
+                ErrorMessage msg ->
+                    msg
+                UnknownError ->
+                    "Encountered unknown error, this usually indicates a bug."
+        exitFailure
 
 mkDefConfiguration :: IO Configuration
 mkDefConfiguration = do