+++ ChangeLog.md
@@ -8,6 +8,13 @@ For latest version of this document see [`ChangeLog.md on GitHub`
 
 * Improved help message
 
+* Split code base into executable and a library, previously there was no
+  library.
+
+* Handle language (translation) preferences in a way that is much closer to
+  [tldr-pages client specification v1.5
+  ](https://github.com/tldr-pages/tldr/blob/v1.5/CLIENT-SPECIFICATION.md)
+
 ### 0.1.0.0
 
 Initial release; not uploaded to Hackage